helm-charts/ray-cluster: Allow multiple workers
We can not set additional workers with the current helm chart.
There are some PRs that aim to fix this but they appear
to be stale https://github.com/ray-project/kuberay/pull/266

We wanted to keep untouched the default worker that is already
included in the helm chart but allow the addition of new workers.

The changes are adding the following logic:

 - Allow to disable the default worker group, if needed, via worker.disabled = true
 - Allow to set additional workers by using the extraWorkers map

The extraWorkers map has a different structure compared to the
default worker because I wanted to add additional flexibility.
Due to that, some variables in the extraWorkers have different
naming (which matches the name of the related CRD attributed)

Note: PR was done to support additional workers in the infrastructure of Metrika.co

Signed-off-by: Christos Kotsis <28815556+ulfox@users.noreply.github.com>